requirements.txt
streamlit
sqlite3
pandas
folium
streamlit-folium
pillow
geocoder
nominatim-api